Move resource loader to where it's actually used.
